+++ CHANGES.md
@@ -2,9 +2,12 @@
 
 ### UX Improvements and new Features
 
+
 ### Fixes
  * Fix handling and formatting of date ranges [#313](https://github.com/CCI-Tools/cate-core/issues/313).
-
+ * The resource switch in the **OPERATION STEP** dialog is only enabled
+   when the drop-down box would contain any **compatible resource**
+   [#310](https://github.com/CCI-Tools/cate-core/issues/310).
 
 ## 0.9.0-dev.4